+++ CHANGELOG.md
@@ -3,7 +3,7 @@
  The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [1.1.0] - 2019-03-07
-- Added support for specific product Id configurations. Example config:
+- Added support for specific product Id configurations and variant products based on base product Id. Example config:
 ```json
 {
   "checkboxValues": {